mode operator related stuff in zom_common::token to its own module
