+++ stage1/zom_common/src/lib.rs
@@ -1,5 +1,6 @@
 //! `zom_common` is related to all things that every Zom crate could want to use, like errors.
 
+pub mod operator;
 pub mod token;
 
 /// Return the current commit hash when this crate was compiled
